add check before delete operation in upload file

+++ sdk/storage/azdatalake/file/client.go
@@ -490,9 +490,11 @@ func (f *Client) uploadFromReader(ctx context.Context, reader io.ReaderAt, actua
 	})
 
 	if err != nil {
-		_, err2 := f.Delete(ctx, nil)
-		if err2 != nil {
-			return exported.ConvertToDFSError(err2)
+		if o.EncryptionContext != nil {
+			_, err2 := f.Delete(ctx, nil)
+			if err2 != nil {
+				return exported.ConvertToDFSError(err2)
+			}
 		}
 		return exported.ConvertToDFSError(err)
 	}
